requirements.txt
absl-py==2.3.1
astunparse==1.6.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.2
fastapi
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
google-pasta==0.2.0
gunicorn==23.0.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.31
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.2.6
opencv-python==4.13.0.90
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
protobuf==6.33.4
pycparser==2.23
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-multipart
pyttsx3==2.99
pytz==2025.2
requests==2.32.5
rich==14.2.0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sounddevice==0.5.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
wrapt==2.0.1
uvicorn